requirements.txt
pydantic
streamlit
python-dotenv
pytest
rich
chromadb
anthropic
deep-translator

